Move tool commands into submodules in zsh/plugins
There are a few external projects that my dotfiles depend on. Instead of
having the user manually add these projects and install the tools, they
are now installed as submodules. This has the added benefit that I can
now depend on their exact location within the filesystem. They are all
installed to ~/.zsh/plugins/.

The current submodules are:
        so-fancy/diff-so-fancy   - human-readable diffs
        jamesiain/zsh-git-prompt - Git status in Zsh prompt