Use Python version of zsh-git-prompt
The technique used by the Haskell version of zsh-git-prompt is
definitely faster, however it isn't worth the hassle of having to
recompile every time I reload the dotfiles configuration from scratch.

There is nothing barring the Python version from using the same
technique that the current Haskell one does. It would be better to add
that support instead.