Add Vim Wiki plugin
The 'vimwiki/vimwiki' is a personal Wiki for Vim. It allows you to
quickly and easily generate and consume a number of linked text files
with their own syntax highlighting.

With Vimwiki you can:
 * organize notes and ideas
 * manage todo-lists
 * write documentation
 * maintain a diary
 * export everything to HTML

Note: I have changed the default syntax to Markdown, as it is more of an
industry standard which allows more options for conversion to other
formats.

Quickstart:
 - Press <leader>ww to go to your wiki's index file. By default, it is
   located in ~/vimwiki/index.wiki
 - You can create links from this file by putting the cursor over a word
   in Normal mode, and pressing <Enter>. This turns that word - or
   visual selection - into a link. Press <Enter> again to follow the
   link to its newly create file.
 - Use <BS> - in Normal mode - to travel backwards up the links.

See https://github.com/vimwiki/vimwiki for more details.