Autocorrect Git command typos
For example, if you type "git cmmit" instead of "git commit" it will
automatically infer the correct command for you.

This only works for very simple typos where there is only one reasonable
resolution. With more complex mistakes - or ones that result in
ambiguity - the default behaviour of failing with an error is
maintained.