Add git churn command
This script - which I borrowed from Gary Bernhardt's config - is used to
evaluate the churn on a set of files. It also sorts them, so that the
ones with the most changes during the specified time period are at the
bottom (for easy access).

This can be used as an indicator of source code (files) which could be
refactored so that they can be more easily used without direct
modification.