Add Vi Normal Mode indicator to right-side prompt
When vi mode is active and the input mode is "Normal", there is now a
reverse video black and yellow NORMAL indication on the right side of
the command line. In all other modes, no indication is shown.

Note: Hitting the <ESC> key is the most common way to enter "Normal"
mode.