Use $(gwhich) and $(gsed) utilities from Homebrew
I am not sure why, but the binaries were renamed in a recent update to
the gnu-which and gnu-sed Homebrew formulae. I assume that it was to
aid in disambiguation.

If it becomes a serious concern, we can always change the name back with
an alias. Currently, it doesn't seem to be that important.