Add zsh-autosuggestions
This plugin provides greyed-out suggestions on the command-line as you
type. Use <End> or <Right-Arrow> to accept then entire line at once. I
have added added a custom bindkey to make <Ctrl-Space> accept up to the
end of the next word in the suggestion.

Note: As zsh-autosuggestions is expected to have been installed using
      'brew', only enable the feature if 'brew' is recognized as a valid
      command by the shell.