Enable custom accept-line override in more places
Previously, the change-prompt-on-accept-line zle widget was only enabled
in the 'main' (or 'viins') keymap. This meant that if the user had
switched into normal mode (AKA 'vicmd' keymap) then the line was not
processed the same way, which leaves some extra crap lying around in the
history section of the terminal.

This change makes both 'viins' and 'vicmd' use the same zle widget to
clean up and time-stamp the command before sending it off to become part
of the history. It makes the shell behave cleaner and more consistently,
regardless of the vi mode which the user is current in.