Start shell with prompt at bottom of the window
This commit forces the command prompt to start at the bottom of the
terminal window. This means - effectively - that it does not ever move
relative to the windows frame. As old command lines are moved into the
history, they slowly creep up over the blank space above.

Since I have started using a transient prompt (in which the prompt is
simplified and marked with a time-stamp in the history) being able to
quickly and easily scan for the currently active prompt is very
valuable. By having it always pinned to the bottom, this is extremely
simple and deterministic.