Discard empty zsh commands w/o advancing the prompt
We now have pinned the active command prompt to the bottom of the
screen, marked the previous commands with a time-stamped transient
prompt, and added a blank line between the commands as a spacer. This
means that the commands are very clearly distinct and easy to follow, so
there is very little reason to hammer out a series of new-lines simply
to space out separate chunks of output.

To that end, I am testing out discarding empty commands completely. I am
curious to see if this becomes annoying in practice. Though I have been
enjoying it for the few minutes that I have tested it so far.